requirements.txt
RPI.BME280
dvb
python-dotenv
requests>=2.31.0
